refactor: restructure codebase into internal packages
- Move API logic to internal/api (client, auth, season, episode, license, manifest, types)
- Move download logic to internal/download (episode, season)
- Move DRM decryption to internal/drm
- Move media manifest/segment parsing to internal/media
- Move muxing to internal/mux
- Move locale mapping to internal/locale
- Update main.go to consume new internal packages
- Remove deprecated flat source files
