+++ internal/media/segment.go
@@ -1,0 +1,185 @@
+package media
+
+import (
+	"bytes"
+	"fmt"
+	"io"
+	"net/http"
+	"os"
+	"strings"
+	"sync"
+	"sync/atomic"
+	"time"
+
+	"github.com/iyear/gowidevine"
+	"github.com/unki2aut/go-mpd"
+)
+
+const maxWorkers = 10
+
+type segmentJob struct {
+	index int
+	url   string
+}
+
+func BuildUrl(base, representationId, file string, partNum *int64) string {
+	if partNum != nil {
+		file = strings.ReplaceAll(file, "$Number$", fmt.Sprintf("%05d", *partNum))
+		file = strings.ReplaceAll(file, "$Number%05d$", fmt.Sprintf("%05d", *partNum))
+	}
+	return base + strings.ReplaceAll(file, "$RepresentationID$", representationId)
+}
+
+func DownloadPart(url string) ([]byte, error) {
+	maxRetries := 5
+	for attempt := 0; attempt < maxRetries; attempt++ {
+		if attempt > 0 {
+			time.Sleep(time.Duration(attempt) * 2 * time.Second)
+		}
+
+		req, err := http.NewRequest(http.MethodGet, url, nil)
+		if err != nil {
+			return nil, err
+		}
+		req.Header.Set("Origin", "https://static.crunchyroll.com")
+		req.Header.Set("Referer", "https://static.crunchyroll.com/")
+		req.Header.Set("User-Agent", "Mozilla/5.0 (X11; Linux x86_64; rv:147.0) Gecko/20100101 Firefox/147.0")
+
+		resp, err := http.DefaultClient.Do(req)
+		if err != nil {
+			if attempt < maxRetries-1 {
+				continue
+			}
+			return nil, fmt.Errorf("failed after %d retries: %w", maxRetries, err)
+		}
+		if resp.StatusCode != 200 {
+			resp.Body.Close()
+			if attempt < maxRetries-1 {
+				continue
+			}
+			return nil, fmt.Errorf("failed after %d retries, status: %d", maxRetries, resp.StatusCode)
+		}
+
+		body, err := io.ReadAll(resp.Body)
+		resp.Body.Close()
+		if err != nil {
+			if attempt < maxRetries-1 {
+				continue
+			}
+			return nil, fmt.Errorf("failed reading body after %d retries: %w", maxRetries, err)
+		}
+		return body, nil
+	}
+	return nil, fmt.Errorf("failed after %d retries", maxRetries)
+}
+
+func getFilename(set *mpd.AdaptationSet) string {
+	if set == nil {
+		f, _ := os.CreateTemp("", "crdl-subs-*.ass")
+		return f.Name()
+	}
+	for _, representation := range set.Representations {
+		if representation.Height != nil {
+			f, _ := os.CreateTemp("", "crdl-video-*.mp4")
+			return f.Name()
+		} else if representation.Bandwidth != nil {
+			f, _ := os.CreateTemp("", "crdl-audio-*.mp3")
+			return f.Name()
+		}
+	}
+	return ""
+}
+
+func DownloadParts(baseUrl, representationId *string, set *mpd.AdaptationSet, keys []*widevine.Key) (string, error) {
+	initUrl := BuildUrl(*baseUrl, *representationId, *set.SegmentTemplate.Initialization, nil)
+	initData, err := DownloadPart(initUrl)
+	if err != nil {
+		return "", err
+	}
+
+	timeline := ExpandTimeline(set.SegmentTemplate.SegmentTimeline.S, 1)
+	total := len(timeline)
+	results := make([][]byte, total)
+	var downloadErr error
+	var errOnce sync.Once
+	var done atomic.Int64
+
+	jobs := make(chan segmentJob, total)
+	var wg sync.WaitGroup
+
+	for w := 0; w < maxWorkers; w++ {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			for job := range jobs {
+				data, err := DownloadPart(job.url)
+				if err != nil {
+					errOnce.Do(func() { downloadErr = err })
+					return
+				}
+				results[job.index] = data
+				count := done.Add(1)
+				fmt.Printf("\rDownloaded %v of %v segments (%v%%)", count, total, (100*count)/int64(total))
+			}
+		}()
+	}
+
+	for i, item := range timeline {
+		url := BuildUrl(*baseUrl, *representationId, *set.SegmentTemplate.Media, &item)
+		jobs <- segmentJob{index: i, url: url}
+	}
+	close(jobs)
+	wg.Wait()
+
+	if downloadErr != nil {
+		return "", downloadErr
+	}
+
+	fmt.Println("\nFinished downloading!")
+
+	var parts []byte
+	parts = append(parts, initData...)
+	for _, data := range results {
+		parts = append(parts, data...)
+	}
+
+	filename := getFilename(set)
+	file, err := os.Create(filename)
+	if err != nil {
+		return "", err
+	}
+	defer file.Close()
+	if err := widevine.DecryptMP4Auto(io.NopCloser(bytes.NewReader(parts)), keys, file); err != nil {
+		return "", fmt.Errorf("widevine.DecryptMP4Auto: %w", err)
+	}
+
+	return filename, nil
+}
+
+func DownloadSubs(url string) (string, error) {
+	req, err := http.NewRequest(http.MethodGet, url, nil)
+	if err != nil {
+		return "", err
+	}
+	req.Header.Set("Origin", "https://static.crunchyroll.com")
+	req.Header.Set("Referer", "https://static.crunchyroll.com/")
+	req.Header.Set("User-Agent", "Mozilla/5.0 (X11; Linux x86_64; rv:147.0) Gecko/20100101 Firefox/147.0")
+
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		return "", err
+	}
+	defer resp.Body.Close()
+
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return "", err
+	}
+
+	filename := getFilename(nil)
+	if err := os.WriteFile(filename, body, 0644); err != nil {
+		return "", err
+	}
+
+	return filename, nil
+}
